2016-08-02 Removed the uncompressed saved sim file to conserve space. Used local instances in GoodsManager.java.
------------------------------------------------------------------------------------------------
I. NEW
------------------------------------------------------------------------------------------------

None.

------------------------------------------------------------------------------------------------------------------
II. IMPROVEMENTS/CHANGES
------------------------------------------------------------------------------------------------------------------

1. Deleted the default_uncompressed file upon compressing it into default.sim by calling .delete() in saveSimulation() in Simulation.java.

2. Used local instances for common config class files in GoodsManager.java.

3. Changed autosaveTimeline to autosaveTimer. 

4. Prevented excessive "Running out of oil" logging statements in Cooking.java.

------------------------------------------------------------------------------------------------------------------
III. FIXES
------------------------------------------------------------------------------------------------------------------

None

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/SimulationConfig.java
@@ -408,24 +408,23 @@ public class SimulationConfig implements Serializable {
 	 * @throws Exception if value is null or empty.
 	 */
 	// 2016-05-02 Added getAutosaveInterval()
-	public double getAutosaveInterval() {
+	public long getAutosaveInterval() {
 		Element root = simulationDoc.getRootElement();
 		Element timeConfig = root.getChild(TIME_CONFIGURATION);
 		Element el = timeConfig.getChild(AUTOSAVE_INTERVAL);
 		String str = el.getAttributeValue(VALUE);
 		
-		double d = 0;
+		long d = 0;
 		
 		if ((str == null) || str.trim().length() == 0)
 			throw new IllegalStateException("autosave_interval must not be blank and must be greater than zero.");
 		else {
 			try {
-		         d = Double.valueOf(str.trim()).doubleValue();
+		         d = (long) Double.valueOf(str.trim()).doubleValue();
 		         //System.out.println("double d = " + d);
 		         
 		         if (d < 1 || d > 1440)
-		 			throw new IllegalStateException("autosave_interval must be between 1 and 1440.");
-		         
+		 			throw new IllegalStateException("autosave_interval must be between 1 and 1440.");		         
 		         
 		      } catch (NumberFormatException nfe) {
 		         System.out.println("NumberFormatException found in autosave_interval : " + nfe.getMessage());
